requirements.txt
Flask
BioPython
